requirements.txt
discord
feedparser
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
mutagen
